Fixed a bug how `finished` how increments completions.
When the test is not reset after its been completed i.e. the user will always
see the alternative they started with, `finished` was incrementing the
completion count of the alternative even if the user had finished the test
before. This lead to situations where altenative's completion count was bigger
its participant count.